refactor(services): move webAiApiCheck to verification module
- relocate service files to logical verification domain
- update imports in runtimeMessages and background tests
